@@ -1,47 +1,49 @@
 {
-  "name": "iron-svelte-client",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "pre-build": "node ./pre-build.js",
-    "pre-build-no-wasm": "node ./pre-build.js no-wasm",
-    "dev": "vite dev",
-    "dev-all": "npm run pre-build && vite dev --host",
-    "dev-no-wasm": "npm run pre-build-no-wasm && vite dev --host",
-    "build": "npm run pre-build && vite build",
-    "build-no-wasm": "npm run pre-build-no-wasm && vite build",
-    "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write ."
-  },
-  "devDependencies": {
-    "@material-design-icons/font": "^0.14.2",
-    "@rollup/plugin-replace": "^5.0.1",
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/adapter-static": "^1.0.0-next.48",
-    "@sveltejs/kit": "next",
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
-    "beercss": "^2.3.0",
-    "cross-env": "7.0.3",
-    "eslint": "^8.16.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "fs-extra": "^11.1.0",
-    "guid-typescript": "^1.0.9",
-    "prettier": "^2.6.2",
-    "prettier-plugin-svelte": "^2.7.0",
-    "rxjs": "^6.6.7",
-    "svelte": "^3.44.0",
-    "svelte-check": "^2.7.1",
-    "svelte-preprocess": "^4.10.6",
-    "tslib": "^2.3.1",
-    "typescript": "^4.7.4",
-    "vite": "^3.2.7",
-    "vite-plugin-top-level-await": "^1.2.2",
-    "vite-plugin-wasm": "^3.1.0"
-  },
-  "type": "module"
+	"name": "iron-svelte-client",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"pre-build": "node ./pre-build.js",
+		"pre-build-no-wasm": "node ./pre-build.js no-wasm",
+		"dev": "vite dev",
+		"dev-all": "npm run pre-build && vite dev --host",
+		"dev-no-wasm": "npm run pre-build-no-wasm && vite dev --host",
+		"build": "npm run pre-build && vite build",
+		"build-no-wasm": "npm run pre-build-no-wasm && vite build",
+		"preview": "vite preview",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --plugin-search-dir . --check . && eslint src/**",
+		"format": "prettier --plugin-search-dir . --write ."
+	},
+	"devDependencies": {
+		"@material-design-icons/font": "^0.14.2",
+		"@rollup/plugin-replace": "^5.0.1",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/adapter-static": "^1.0.0-next.48",
+		"@sveltejs/kit": "next",
+		"@typescript-eslint/eslint-plugin": "^5.27.0",
+		"@typescript-eslint/parser": "^5.62.0",
+		"beercss": "^2.3.0",
+		"cross-env": "7.0.3",
+		"eslint": "^8.52.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-html": "^7.1.0",
+		"eslint-plugin-svelte": "^2.34.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"fs-extra": "^11.1.0",
+		"guid-typescript": "^1.0.9",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"rxjs": "^6.6.7",
+		"svelte": "^3.59.2",
+		"svelte-check": "^2.7.1",
+		"svelte-preprocess": "^4.10.6",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.4",
+		"vite": "^3.2.7",
+		"vite-plugin-top-level-await": "^1.2.2",
+		"vite-plugin-wasm": "^3.1.0"
+	},
+	"type": "module"
 }
