@@ -1,10 +1,10 @@
 ﻿export enum LockKey {
-    CAPS_LOCK = "CapsLock",
-    NUM_LOCK = "NumLock",
-    SCROLL_LOCK = "ScrollLock",
-    KANA_MODE = "KanaMode",
-    "CapsLock" = CAPS_LOCK,
-    "ScrollLock" = SCROLL_LOCK,
-    "NumLock" = NUM_LOCK,
-    "KanaMode" = KANA_MODE
+	CAPS_LOCK = 'CapsLock',
+	NUM_LOCK = 'NumLock',
+	SCROLL_LOCK = 'ScrollLock',
+	KANA_MODE = 'KanaMode',
+	'CapsLock' = CAPS_LOCK,
+	'ScrollLock' = SCROLL_LOCK,
+	'NumLock' = NUM_LOCK,
+	'KanaMode' = KANA_MODE
 }
