@@ -1,8 +1,8 @@
 ﻿export enum LogType {
-    OFF = "OFF",
-    ERROR = "ERROR",
-    WARN = "WARN",
-    INFO = "INFO",
-    DEBUG = "DEBUG",
-    TRACE = "TRACE"
+	OFF = 'OFF',
+	ERROR = 'ERROR',
+	WARN = 'WARN',
+	INFO = 'INFO',
+	DEBUG = 'DEBUG',
+	TRACE = 'TRACE'
 }
