@@ -1,14 +1,14 @@
 ﻿export enum ModifierKey {
-    CTRL_LEFT = "ControlLeft",
-    SHIFT_LEFT = "ShiftLeft",
-    SHIFT_RIGHT = "ShiftRight",
-    ALT_LEFT = "AltLeft",
-    CTRL_RIGHT = "ControlRight",
-    ALT_RIGHT = "AltRight",
-    "ControlLeft" = CTRL_LEFT,
-    "ShiftLeft" = SHIFT_LEFT,
-    "ShiftRight" = SHIFT_RIGHT,
-    "AltLeft" = ALT_LEFT,
-    "ControlRight" = CTRL_RIGHT,
-    "AltRight" = ALT_RIGHT,
+	CTRL_LEFT = 'ControlLeft',
+	SHIFT_LEFT = 'ShiftLeft',
+	SHIFT_RIGHT = 'ShiftRight',
+	ALT_LEFT = 'AltLeft',
+	CTRL_RIGHT = 'ControlRight',
+	ALT_RIGHT = 'AltRight',
+	'ControlLeft' = CTRL_LEFT,
+	'ShiftLeft' = SHIFT_LEFT,
+	'ShiftRight' = SHIFT_RIGHT,
+	'AltLeft' = ALT_LEFT,
+	'ControlRight' = CTRL_RIGHT,
+	'AltRight' = ALT_RIGHT
 }
