@@ -1,7 +1,7 @@
 ﻿export enum MouseButton {
-    LEFT = 0,
-    MIDDLE = 1,
-    RIGHT = 2,
-    BROWSER_BACK = 3,
-    BROWSER_FORWARD = 4
+	LEFT = 0,
+	MIDDLE = 1,
+	RIGHT = 2,
+	BROWSER_BACK = 3,
+	BROWSER_FORWARD = 4
 }
