@@ -1,5 +1,5 @@
 ﻿export enum OS {
-    WINDOWS = 'windows',
-    LINUX = 'linux',
-    MACOS = 'macos'
+	WINDOWS = 'windows',
+	LINUX = 'linux',
+	MACOS = 'macos'
 }
