@@ -1,5 +1,5 @@
 ﻿export enum ScreenScale {
-    Fit = 1,
-    Full = 2,
-    Real = 3
+	Fit = 1,
+	Full = 2,
+	Real = 3
 }
