@@ -1,5 +1,5 @@
 ﻿export enum SessionEventType {
-    STARTED,
-    TERMINATED,
-    ERROR
+	STARTED,
+	TERMINATED,
+	ERROR
 }
