@@ -1,4 +1,4 @@
 ﻿export enum SpecialCombination {
-    CTRL_ALT_DEL,
-    META
+	CTRL_ALT_DEL,
+	META
 }
