@@ -1,4 +1,4 @@
 ﻿export interface DesktopSize {
-    width: number,
-    height: number
+	width: number;
+	height: number;
 }
