@@ -1,7 +1,7 @@
-﻿import type {DesktopSize} from './DesktopSize';
+﻿import type { DesktopSize } from './DesktopSize';
 
 export interface NewSessionInfo {
-    session_id: number,
-    websocket_port: number,
-    initial_desktop_size: DesktopSize,
+	session_id: number;
+	websocket_port: number;
+	initial_desktop_size: DesktopSize;
 }
