@@ -1,6 +1,6 @@
-﻿import type {DesktopSize} from './DesktopSize';
+﻿import type { DesktopSize } from './DesktopSize';
 
 export interface ResizeEvent {
-    session_id: number,
-    desktop_size: DesktopSize,
+	session_id: number;
+	desktop_size: DesktopSize;
 }
