@@ -1,19 +1,27 @@
-﻿import type {ScreenScale} from '../enums/ScreenScale';
-import type {Observable} from 'rxjs';
-import type {NewSessionInfo} from './NewSessionInfo';
+﻿import type { ScreenScale } from '../enums/ScreenScale';
+import type { Observable } from 'rxjs';
+import type { NewSessionInfo } from './NewSessionInfo';
+import type { SessionEvent } from './session-event';
 
 export interface UserInteraction {
-    setVisibility(state: boolean);
+	setVisibility(state: boolean);
 
-    setScale(scale: ScreenScale);
+	setScale(scale: ScreenScale);
 
-    connect(username: string, password: string, destination: string, proxyAddress: string, serverDomain: string, authToken: string): Observable<NewSessionInfo>;
+	connect(
+		username,
+		password: string,
+		destination: string,
+		proxyAddress: string,
+		serverDomain: string,
+		authToken: string
+	): Observable<NewSessionInfo>;
 
-    ctrlAltDel();
+	ctrlAltDel();
 
-    metaKey();
-    
-    shutdown();
+	metaKey();
 
-    sessionListener: Observable<any>;
+	shutdown();
+
+	sessionListener: Observable<SessionEvent>;
 }
