@@ -1,9 +1,9 @@
-import IronRemoteGui from './iron-remote-gui.svelte';
+export * as default from './iron-remote-gui.svelte';
 
-export type {UserInteraction} from './interfaces/UserInteraction';
-export type {ResizeEvent} from './interfaces/ResizeEvent';
-export type {NewSessionInfo} from './interfaces/NewSessionInfo';
-export type {ServerRect} from './interfaces/ServerRect';
-export type {DesktopSize} from './interfaces/DesktopSize';
-export type {SessionEvent} from './interfaces/session-event';
-export type {SessionEventType} from './enums/SessionEventType';
+export type { UserInteraction } from './interfaces/UserInteraction';
+export type { ResizeEvent } from './interfaces/ResizeEvent';
+export type { NewSessionInfo } from './interfaces/NewSessionInfo';
+export type { ServerRect } from './interfaces/ServerRect';
+export type { DesktopSize } from './interfaces/DesktopSize';
+export type { SessionEvent } from './interfaces/session-event';
+export type { SessionEventType } from './enums/SessionEventType';
