@@ -1,10 +1,12 @@
-import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
-export default {
-  // Consult https://svelte.dev/docs#compile-time-svelte-preprocess
-  // for more information about preprocessors
-  compilerOptions: {
-    customElement: true
-  },
-  preprocess: vitePreprocess(),
-}
+const config = {
+	// Consult https://svelte.dev/docs#compile-time-svelte-preprocess
+	// for more information about preprocessors
+	compilerOptions: {
+		customElement: true
+	},
+	preprocess: vitePreprocess()
+};
+
+export default config;
