@@ -1,7 +1,7 @@
 export class Rect {
-  top!: number;
-  left!: number;
-  width!: number;
-  height!: number;
-  buffer!: ArrayBuffer
+	top!: number;
+	left!: number;
+	width!: number;
+	height!: number;
+	buffer!: ArrayBuffer;
 }
