@@ -1,4 +1,4 @@
 <script>
-    import { goto } from '$app/navigation';
-    goto("/session");
+	import { goto } from '$app/navigation';
+	goto('/session');
 </script>
