@@ -1,15 +1,13 @@
 <script lang="ts">
-    import Login from '$lib/login/login.svelte';
-    import RemoteScreen from '$lib/remote-screen/remote-screen.svelte';
-    import Message from '$lib/messages/message.svelte';
-    import {showLogin} from "$lib/login/login-store.ts";
-
+	import Login from '$lib/login/login.svelte';
+	import RemoteScreen from '$lib/remote-screen/remote-screen.svelte';
+	import Message from '$lib/messages/message.svelte';
+	import { showLogin } from '$lib/login/login-store.ts';
 </script>
 
 {#if $showLogin}
-    <Login/>
+	<Login />
 {/if}
-<RemoteScreen/>
+<RemoteScreen />
 
-<Message></Message>
-
+<Message />
