@@ -14,7 +14,7 @@ const config = {
 			fallback: null,
 			precompress: false,
 			strict: true
-	  })
+		})
 	}
 };
 
