@@ -5,7 +5,7 @@ import topLevelAwait from 'vite-plugin-top-level-await';
 
 const config: UserConfig = {
 	mode: process.env.MODE || 'development',
-	plugins: [sveltekit(), wasm(), topLevelAwait()],
+	plugins: [sveltekit(), wasm(), topLevelAwait()]
 };
 
 export default config;
